Move kernel headers into dedicated directory
