+++ src-headers/qspinlock.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../spinlock.h"
+#include "spinlock.h"
 
 void qspin_lock(struct spinlock *lk);
 void qspin_unlock(struct spinlock *lk);
