+++ tests/test_lockstress.py
@@ -101,6 +101,7 @@ def compile_and_run():
             "gcc","-std=c11","-pthread","-DSPINLOCK_NO_STUBS",
             "-I", str(ROOT),
             "-I", str(ROOT/"src-headers"),
+            "-I", str(ROOT/"src-kernel/include"),
             str(src),
             "-o", str(exe)
         ])
