+++ tests/test_zone.py
@@ -49,7 +49,8 @@ def compile_and_run(body):
         exe = pathlib.Path(td)/"test"
         src.write_text(C_CODE_TEMPLATE % body)
         # headers expected by zone.c
-        (pathlib.Path(td)/"spinlock.h").write_text('#include "src-headers/spinlock.h"\n')
+        (pathlib.Path(td)/"spinlock.h").write_text(
+            '#include "src-headers/libos/spinlock.h"\n')
         (pathlib.Path(td)/"defs.h").write_text("")
         (pathlib.Path(td)/"mmu.h").write_text('#include "src-headers/types.h"\n#include "src-headers/mmu.h"\n')
         (pathlib.Path(td)/"memlayout.h").write_text('#include "src-headers/memlayout.h"\n')
